refactor: move git logic to separate file
